refactor to BaseComponentTestCase
