+++ tests/unit/connector/base.py
@@ -15,11 +15,12 @@ from logprep.abc.connector import Connector
 from logprep.abc.input import Input
 from logprep.abc.output import Output
 from logprep.factory import Factory
-from logprep.util.helper import camel_to_snake
 from logprep.util.time_measurement import TimeMeasurement
 
+from tests.unit.component.base import BaseCompontentTestCase
 
-class BaseConnectorTestCase(ABC):
+
+class BaseConnectorTestCase(BaseCompontentTestCase):
     CONFIG: dict = {}
     object: Connector = None
     logger = getLogger()
@@ -30,24 +31,6 @@ class BaseConnectorTestCase(ABC):
 
     def test_is_a_connector_implementation(self):
         assert isinstance(self.object, Connector)
-
-    def test_uses_python_slots(self):
-        assert isinstance(self.object.__slots__, Iterable)
-
-    def test_describe(self):
-        describe_string = self.object.describe()
-        expected_base_description = f"{self.object.__class__.__name__} (Test Instance Name)"
-        assert describe_string.startswith(expected_base_description)
-
-    def test_snake_type(self):
-        assert str(self.object) == camel_to_snake(self.object.__class__.__name__)
-
-    def test_schedules_tasks(self):
-        mock_task = mock.MagicMock()
-        self.object._schedule_task(task=mock_task, seconds=1)
-        with mock.patch("schedule.Job.should_run", return_value=True):
-            self.object.run_pending_tasks()
-        mock_task.assert_called()
 
 
 class BaseInputTestCase(BaseConnectorTestCase):
